api/requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
pytest==7.4.0
pymongo
python-dotenv
flake8
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0